refactor: collapse grant expand skip and split event tests
Invert the QueryGrants session expand gate once instead of doubling
!userCaller. Move assignment-project event assertions out of
grant_test.go (now 891 lines). Share SeedProjectViewer for session
integration footholds.
