+++ internal/api/integration_test/grant_test.go
@@ -623,17 +623,7 @@ func TestGrantSessionCaller(t *testing.T) {
 	require.NoError(t, err)
 
 	operatorID := harness.CreateUserWithTeam(t, platform.ID)
-	// MVP catalog: assigned viewer closes to editor/admin Checks (seed until #420).
-	operatorAsgn := &domain.AuthzAssignment{
-		ProjectID:     project.ID,
-		CatalogID:     domain.SystemCatalogID,
-		PrincipalType: domain.AuthzPrincipalTypeUser,
-		PrincipalID:   operatorID,
-		ObjectType:    "project",
-		Relation:      "viewer",
-	}
-	operatorAsgn.ApplyScope(domain.NewProjectAssignmentScope())
-	require.NoError(t, harness.EnsureServiceDB(t).Statements().CreateAuthzAssignment(t.Context(), operatorAsgn))
+	harness.SeedProjectViewer(t, project.ID, operatorID)
 
 	client, err := helpers.NewApiClient(harness.EnsureTestServer(t).URL)
 	require.NoError(t, err)
